Aggregation methods

FuzzyLogic.MaxAggregator — Type

struct MaxAggregator <: FuzzyLogic.AbstractAggregator

Aggregator that combines fuzzy rules output by taking their maximum.

FuzzyLogic.ProbSumAggregator — Type

struct ProbSumAggregator <: FuzzyLogic.AbstractAggregator

Aggregator that combines fuzzy rules output by taking their probabilistic sum. See also ProbSumOr.