Defuzzification methods

FuzzyLogic.BisectorDefuzzifier — Type

struct BisectorDefuzzifier <: FuzzyLogic.AbstractDefuzzifier

Bisector defuzzifier. Given the aggregated output function $f$ and the output variable domain $[a, b]$ the defuzzified output is the value $t ∈ [a, b]$ that divides the area under $f$ into two equal parts. That is

\[∫_a^tf(x)\textrm{d}x = ∫_t^af(x)\textrm{d}x.\]

Algorithm

The domain is partitioned into N equal subintervals. For each subinterval endpoint, the left and right area are approximated using the trapezoidal rule. The end point leading to the best approximation is the final result.

FuzzyLogic.CentroidDefuzzifier — Type

struct CentroidDefuzzifier <: FuzzyLogic.AbstractDefuzzifier

Centroid defuzzifier. Given the aggregated output function $f$ and the output variable domain $[a, b]$ the defuzzified output is the centroid computed as

\[\frac{∫_a^bxf(x)\textrm{d}x}{∫_a^bf(x)\textrm{d}x}.\]

Algorithm

The integrals are computed numerically using the trapezoidal rule.